+++ src/solvers/cc/diis_cc_es_class.F90
@@ -137,11 +137,12 @@ contains
       logical :: crop
       integer :: diis_dimension, state
 !
-      this%wf => wf 
+      this%wf => wf
 !
-      this%timer = timings(trim(convert_to_uppercase(this%wf%name_)) // &
-                     ' excited state (' // trim(transformation) //')')
-      call this%timer%turn_on()
+      this%iteration_timer = timings('DIIS CC ES solver iteration')
+      this%total_timer = timings('DIIS CC ES solver (' // &
+                                 trim(transformation) //')')
+      call this%total_timer%turn_on()
 !
 !     Set printables
 !
@@ -149,9 +150,9 @@ contains
       this%tag    = 'DIIS'
 !
       this%description1 = 'A DIIS solver that solves for the lowest eigenvalues and &
-                           & the right eigenvectors of the Jacobian matrix, A. The eigenvalue &
-                           & problem is solved by DIIS extrapolation of residuals for each &
-                           & eigenvector until the convergence criteria are met.'
+                           &the eigenvectors of the Jacobian matrix, A. The eigenvalue &
+                           &problem is solved by DIIS extrapolation of residuals for each &
+                           &eigenvector until the convergence criteria are met.'
 !
       this%description2 = 'More on the DIIS algorithm can be found in &
                            &P. Pulay, Chemical Physics Letters, 73(2), 393-398 (1980).'
@@ -283,9 +284,6 @@ contains
 !
       real(dp) :: ddot
 !
-      type(timings) :: iteration_time
-      character(len=30) :: timer_name
-!
 !     Prepare wavefunction for excited state, and possibly do Davidson preconvergence
 !
       call this%prepare_wf_for_excited_state()
@@ -338,11 +336,9 @@ contains
 !
       do while (.not. all(converged) .and. (iteration .le. this%max_iterations))
 !
-         iteration = iteration + 1
+         call this%iteration_timer%turn_on()
 !
-         write(timer_name, '(a,i0)') 'DIIS: CC ES iteration ', iteration
-         iteration_time = timings(trim(timer_name), pl="n")
-         call iteration_time%turn_on()
+         iteration = iteration + 1
 !
          call output%printf('n', 'Iteration: (i18)', ints=[iteration], fs='(/t3,a)')
 !
@@ -424,8 +420,8 @@ contains
 !
          call output%print_separator('n', 47, '-')
 !
-         call iteration_time%turn_off()
-         call iteration_time%reset()
+         call this%iteration_timer%turn_off()
+         call this%iteration_timer%reset()
 !
       enddo
 !